refactor(reporter-v2): rename entrypoint to article_generator
Match the module name to its role as the product composition root that
wires tools, completion, and seeded artifacts into Runner.
