+++ reporter_v2/__init__.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 from reporter_v2.config import BiasProfile, ReportConfig, TimeRange, ToneControls
-from reporter_v2.runner.entrypoint import generate_article
+from reporter_v2.runner.article_generator import generate_article
 from reporter_v2.runner.runner import Runner
 from reporter_v2.runner.schemas import ArticleOutput
 from reporter_v2.runner.state import ProcedureHistoryMode, RunnerConfig
