+++ reporter_v2/app/runner.py
@@ -16,7 +16,7 @@ from datalayer.sleeper_data import SleeperConfig, SleeperLeagueData
 from reporter_memory.context_store import ContextStore
 from reporter_v2.config import BiasProfile, ReportConfig, TimeRange, ToneControls
 from reporter_v2.runner.completion import CompletionSettings, RetryPolicy
-from reporter_v2.runner.entrypoint import generate_article
+from reporter_v2.runner.article_generator import generate_article
 from reporter_v2.runner.state import ProcedureHistoryMode, RunnerConfig
