+++ reporter_v2/tests/test_integration.py
@@ -1,4 +1,4 @@
-"""Integration tests for the reporter v2 entrypoint."""
+"""Integration tests for the reporter v2 article generator."""
 
 from __future__ import annotations
 
@@ -11,7 +11,7 @@ from typing import Any
 from reporter_memory.context_store import ContextStore
 from reporter_v2.config import ReportConfig, TimeRange
 from reporter_v2.runner.completion import CompletionSettings
-from reporter_v2.runner.entrypoint import _persist_brief_facts, generate_article
+from reporter_v2.runner.article_generator import _persist_brief_facts, generate_article
 from reporter_v2.runner.models import ToolCall
 from reporter_v2.runner.schemas import Fact, ReportBrief, Storyline
