+++ reporter_v2/workflows.py
@@ -11,7 +11,7 @@ from datalayer.sleeper_data import SleeperLeagueData
 from reporter_memory.context_store import ContextStore
 from reporter_v2.config import BiasProfile, ReportConfig, TimeRange, ToneControls
 from reporter_v2.runner.completion import CompletionSettings
-from reporter_v2.runner.entrypoint import generate_article
+from reporter_v2.runner.article_generator import generate_article
 from reporter_v2.runner.schemas import ArticleOutput
 from reporter_v2.runner.state import RunnerConfig
